Move oh-icon to system library
