+++ bundles/org.openhab.ui/web/src/js/app.js
@@ -3,7 +3,7 @@ import Vue from 'vue'
 
 // Import globally registered components
 import SitemapWidgetGeneric from '../components/sitemap/widget-generic.vue'
-import OHIconComponent from '../components/oh-icon.vue'
+import OHIconComponent from '../components/widgets/system/oh-icon.vue'
 import ModelTreeviewItem from '../components/model/treeview-item.vue'
 import SitemapTreeviewItem from '../components/pagedesigner/sitemap/treeview-item.vue'
 import EmptyStatePlaceholder from '../components/empty-state-placeholder.vue'
